move fluid generator to phi
