+++ paddle/fluid/distributed/ps/table/depends/initializers.h
@@ -22,8 +22,8 @@
 #include <vector>
 
 #include "gflags/gflags.h"
-#include "paddle/fluid/framework/generator.h"
 #include "paddle/fluid/operators/truncated_gaussian_random_op.h"
+#include "paddle/phi/core/generator.h"
 
 namespace paddle {
 namespace distributed {
@@ -64,7 +64,7 @@ class UniformInitializer : public Initializer {
     max_ = std::stof(attrs[3]);
 
     dist_ = std::uniform_real_distribution<float>(min_, max_);
-    random_engine_ = framework::GetCPURandomEngine(seed_);
+    random_engine_ = phi::GetCPURandomEngine(seed_);
   }
 
   float GetValue() override { return dist_(*random_engine_); }
@@ -90,7 +90,7 @@ class GaussianInitializer : public Initializer {
     mean_ = std::stof(attrs[2]);
     std_ = std::stof(attrs[3]);
 
-    random_engine_ = framework::GetCPURandomEngine(seed_);
+    random_engine_ = phi::GetCPURandomEngine(seed_);
 
     dist_ = std::normal_distribution<float>(mean_, std_);
   }
@@ -120,7 +120,7 @@ class TruncatedGaussianInitializer : public Initializer {
 
     std::uniform_real_distribution<float> dist_(
         std::numeric_limits<float>::min(), 1.0);
-    random_engine_ = framework::GetCPURandomEngine(seed_);
+    random_engine_ = phi::GetCPURandomEngine(seed_);
   }
 
   float GetValue() override {
