+++ paddle/fluid/operators/seed_op.h
@@ -13,9 +13,9 @@
 // limitations under the License.
 #pragma once
 
-#include "paddle/fluid/framework/generator.h"
 #include "paddle/fluid/framework/op_registry.h"
 #include "paddle/fluid/framework/op_version_registry.h"
+#include "paddle/phi/core/generator.h"
 
 namespace paddle {
 namespace operators {
@@ -36,7 +36,7 @@ static int get_seed(const framework::ExecutionContext& context) {
     }
   } else {
     std::string name = context.Attr<std::string>("rng_name");
-    auto rng = framework::GetRandomSeedGenerator(name);
+    auto rng = phi::GetRandomSeedGenerator(name);
     do {  // NOTE(wangxi): cpu dropout will use random seed if seed == 0
       seed = static_cast<int>(rng->Random64());
     } while (seed == 0);
