+++ paddle/fluid/operators/truncated_gaussian_random_op.cc
@@ -18,9 +18,9 @@ limitations under the License. */
 #include <random>
 #include <vector>
 
-#include "paddle/fluid/framework/generator.h"
 #include "paddle/fluid/framework/infershape_utils.h"
 #include "paddle/fluid/framework/op_registry.h"
+#include "paddle/phi/core/generator.h"
 #include "paddle/phi/infermeta/nullary.h"
 
 namespace paddle {
