+++ paddle/fluid/operators/truncated_gaussian_random_op_npu.cc
@@ -90,7 +90,7 @@ class NPUTruncatedGaussianRandomKernel : public framework::OpKernel<T> {
     int64_t size = tensor->numel();
 
     unsigned int seed = static_cast<unsigned int>(context.Attr<int>("seed"));
-    auto engine = framework::GetCPURandomEngine(seed);
+    auto engine = phi::GetCPURandomEngine(seed);
     for (int64_t i = 0; i < size; ++i) {
       cpu_data[i] = truncated_normal(dist(*engine));
     }
